+++ src/ts/AdUnits/AbstractAdUnit.ts
@@ -2,6 +2,7 @@ import { Placement } from 'Models/Placement';
 import { Campaign } from 'Models/Campaign';
 import { Observable0 } from 'Utilities/Observable';
 import { NativeBridge } from 'Native/NativeBridge';
+import { AdUnit } from 'Utilities/AdUnit';
 
 export abstract class AbstractAdUnit {
 
@@ -20,11 +21,13 @@ export abstract class AbstractAdUnit {
     public onClose: Observable0 = new Observable0();
 
     protected _nativeBridge: NativeBridge;
+    protected _adUnit: AdUnit;
     protected _placement: Placement;
     protected _campaign: Campaign;
 
-    constructor(nativeBridge: NativeBridge, placement: Placement, campaign: Campaign) {
+    constructor(nativeBridge: NativeBridge, adUnit: AdUnit, placement: Placement, campaign: Campaign) {
         this._nativeBridge = nativeBridge;
+        this._adUnit = adUnit;
         this._placement = placement;
         this._campaign = campaign;
     }
@@ -34,6 +37,8 @@ export abstract class AbstractAdUnit {
     public abstract hide(): Promise<void>;
 
     public abstract isShowing(): boolean;
+
+    public abstract description(): string;
 
     public getPlacement(): Placement {
         return this._placement;