+++ src/ts/EventHandlers/VideoEventHandlers.ts
@@ -14,6 +14,7 @@ import { DiagnosticError } from 'Errors/DiagnosticError';
 import { VideoAdUnit } from 'AdUnits/VideoAdUnit';
 import { PerformanceCampaign } from 'Models/PerformanceCampaign';
 import { VastCampaign } from 'Models/Vast/VastCampaign';
+import { Overlay } from 'Views/Overlay';
 
 export class VideoEventHandlers {
 
@@ -57,8 +58,8 @@ export class VideoEventHandlers {
             if(adUnit.getVideoAdUnitController().getVideoPosition() > 0) {
                 overlay.setVideoProgress(adUnit.getVideoAdUnitController().getVideoPosition());
             }
-            if(adUnit.getPlacement().allowSkip()) {
-                overlay.setSkipVisible(true);
+            if(overlay instanceof Overlay && adUnit.getPlacement().allowSkip()) {
+                (<Overlay>overlay).setSkipVisible(true);
             }
             overlay.setMuteEnabled(true);
             overlay.setVideoDurationEnabled(true);