+++ src/ts/Views/EndScreen.ts
@@ -89,10 +89,6 @@ export class EndScreen extends View {
     }
 
     private getEndscreenAlt(campaign: PerformanceCampaign) {
-        const abGroup = campaign.getAbGroup();
-        if(abGroup !== 8 && abGroup !== 9) {
-            return 'xmas';
-        }
         return undefined;
     }